tg_bot/requirements.txt
django-environ>=0.11.2
python-telegram-bot>=21.2
requests>=2.31.0
